refactor: extract shared litert-shared/ module for the on-device demos
Move litert-engine.ts, resume-store.ts, eval-hud.ts, and the demo chrome
CSS (litert-slides.css -> litert-chrome.css) out of litert-slides/ into a
litert-shared/ directory, and extract the copy-pasted model-picker wiring
(dropdown fill, load button, WebGPU gate, status line) from all four
demo main.ts files into wireModelLoader(). Per-demo deltas stay at the
call site: the ready-status hint and litert-intake's E4B default. All
demos now share one model registry, loader, cache, and HUD by
construction.
