+++ apps/web/src/litert-paint/main.ts
@@ -3,7 +3,7 @@
 // is Gemma 4 running in the browser via LiteRT-LM (WebGPU). No proxy, no hosted
 // runtime, no API key. The paint host, bridge, tools, and Win98 theme are
 // reused verbatim from ../webmcp-paint; the only difference is the engine
-// behind Persona (see ../litert-slides/litert-engine.ts).
+// behind Persona (see ../litert-shared/litert-engine.ts).
 //
 // Prompting follows Google's Gemma 4 formatting guide
 // (https://ai.google.dev/gemma/docs/core/prompt-formatting-gemma4). Most of the
@@ -24,9 +24,9 @@
 //     balloon the small context window this demo runs in.
 import "@runtypelabs/persona/widget.css";
 import "../webmcp-paint/paint.css";
-// litert-slides.css is imported for the shared eval-HUD styles (.lr-hud…);
+// litert-chrome.css is imported for the shared toolbar + eval-HUD styles (.lr-*);
 // litert-paint.css restyles the HUD + model controls to match the Win98 shell.
-import "../litert-slides/litert-slides.css";
+import "../litert-shared/litert-chrome.css";
 import "./litert-paint.css";
 
 import {
@@ -42,13 +42,12 @@ import { initializeWebMCPPolyfill } from "@mcp-b/webmcp-polyfill";
 import { mountJsPaint, type PaintBridge } from "../webmcp-paint/jspaint-host";
 import { paintTheme } from "../webmcp-paint/theme";
 import { APPROVAL_REQUIRED_TOOL_NAMES, setupPaintTools } from "../webmcp-paint/tools";
-import { createEvalHud } from "../litert-slides/eval-hud";
+import { createEvalHud } from "../litert-shared/eval-hud";
 import {
-  MODELS,
   type LiteRtPersonaEngine,
-  type ModelId,
   createLiteRtPersonaEngine,
-} from "../litert-slides/litert-engine";
+} from "../litert-shared/litert-engine";
+import { wireModelLoader } from "../litert-shared/model-loader";
 
 initializeWebMCPPolyfill();
 
@@ -123,61 +122,12 @@ window.personaLiteRtEngine = engine;
 
 // ---- Model picker (E2B default, swap to E4B) ------------------------------
 
-const modelSelect = document.querySelector<HTMLSelectElement>("#lr-model-select");
-const loadButton = document.querySelector<HTMLButtonElement>("#lr-load-button");
-const statusEl = document.querySelector<HTMLElement>("#lr-status");
-const webgpuWarning = document.querySelector<HTMLElement>("#lr-webgpu-warning");
-
-const webgpuSupported = typeof navigator !== "undefined" && "gpu" in navigator;
-
-if (modelSelect) {
-  for (const id of Object.keys(MODELS) as ModelId[]) {
-    const info = MODELS[id];
-    const option = document.createElement("option");
-    option.value = id;
-    option.textContent = `${info.label} (${info.approxSize})`;
-    option.title = info.blurb;
-    modelSelect.appendChild(option);
-  }
-  modelSelect.value = "e2b";
-}
-
-const setStatus = (text: string): void => {
-  if (statusEl) statusEl.textContent = text;
-};
-
-async function loadSelectedModel(): Promise<void> {
-  if (!modelSelect || !loadButton) return;
-  const modelId = modelSelect.value as ModelId;
-  loadButton.disabled = true;
-  modelSelect.disabled = true;
-  // loadModel downloads the weights, then warms up the GPU with a throwaway
-  // generation so the first real prompt is fast — that warm-up can take a few
-  // minutes on first run. Set the expectation up front.
-  setStatus(
-    `Loading ${MODELS[modelId].label}… the first load downloads ${MODELS[modelId].approxSize} (cached for next time), then warms up the GPU — the first run can take a few minutes.`,
-  );
-  try {
-    await engine.loadModel(modelId);
-    setStatus(`${MODELS[modelId].label} ready (warmed up) — ask Paint Pal for a drawing.`);
-    loadButton.textContent = "Reload";
-  } catch (err) {
-    setStatus(`Load failed: ${err instanceof Error ? err.message : String(err)}`);
-  } finally {
-    loadButton.disabled = false;
-    modelSelect.disabled = false;
-  }
-}
-
-if (!webgpuSupported) {
-  webgpuWarning?.removeAttribute("hidden");
-  if (loadButton) loadButton.disabled = true;
-  if (modelSelect) modelSelect.disabled = true;
-  setStatus("WebGPU unavailable");
-} else {
-  loadButton?.addEventListener("click", () => void loadSelectedModel());
-  setStatus("Pick a model and press Load to start (runs fully on-device).");
-}
+// Shared wiring (E2B default); keep the setStatus handle for the jspaint
+// bridge failure message below.
+const { setStatus } = wireModelLoader({
+  engine,
+  readyHint: "ask Paint Pal for a drawing.",
+});
 
 // ---- Persona widget --------------------------------------------------------
 
@@ -194,7 +144,7 @@ function mountWidget(): void {
       ...DEFAULT_WIDGET_CONFIG,
       // The engine answers this path from the in-browser model (no network).
       apiUrl: API_PATH,
-      // The fetch patch in ../litert-slides/litert-engine handles the fake
+      // The fetch patch in ../litert-shared/litert-engine handles the fake
       // dispatch/resume routes. This early dispatch guard is demo UX: the
       // widget only paints an assistant fallback bubble when dispatch rejects
       // before an SSE stream starts, so fail fast here when the user chats
